AB#23401 updates to tinyMCE styles handling and add notification module bundle
