+++ applications/Unity.GrantManager/modules/Unity.Notifications/src/Unity.Notifications.Web/Bundling/NotificationsBundles.cs
@@ -1,0 +1,9 @@
+﻿namespace Unity.Notifications.Web.Bundling;
+
+public static class NotificationsBundles
+{
+    public static class Styles
+    {
+        public const string Notifications = "Notifications";
+    }
+}
