+++ applications/Unity.GrantManager/modules/Unity.Notifications/src/Unity.Notifications.Web/Views/Settings/NotificationsSettingGroup/Default.cshtml
@@ -1,4 +1,5 @@
-﻿@model Unity.Notifications.Web.Views.Settings.NotificationsSettingGroup.NotificationsSettingViewModel
+﻿@using Unity.Notifications.Web.Bundling
+@model Unity.Notifications.Web.Views.Settings.NotificationsSettingGroup.NotificationsSettingViewModel
 
 <abp-script src="/Views/Settings/NotificationsSettingGroup/Default.js" />
 <abp-style src="/Views/Settings/NotificationsSettingGroup/Default.css" />
