+++ applications/Unity.GrantManager/modules/Unity.Notifications/src/Unity.Notifications.Web/Views/Settings/NotificationsSettingGroup/NotificationsSettingViewComponent.cs
@@ -1,14 +1,17 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Volo.Abp.AspNetCore.Mvc;
+using Volo.Abp.AspNetCore.Mvc.UI.Bundling;
 using Volo.Abp.AspNetCore.Mvc.UI.Widgets;
+using Volo.Abp.Bundling;
 using Volo.Abp.Settings;
 
 namespace Unity.Notifications.Web.Views.Settings.NotificationsSettingGroup;
 
 [Widget(
-    ScriptTypes = new[] { typeof(NotificationsSettingScriptBundleContributor) },
-    StyleTypes = new[] { typeof(NotificationsSettingStyleBundleContributor) },
+    ScriptTypes = [typeof(NotificationsSettingScriptBundleContributor)],
+    StyleTypes = [typeof(NotificationsSettingStyleBundleContributor)],
     AutoInitialize = true
 )]
 public class NotificationsSettingViewComponent(ISettingProvider settingProvider) : AbpViewComponent
@@ -26,6 +29,25 @@ public class NotificationsSettingViewComponent(ISettingProvider settingProvider)
         };
 
         return View("~/Views/Settings/NotificationsSettingGroup/Default.cshtml", model);
+    }
+
+    public class NotificationsSettingScriptBundleContributor : BundleContributor
+    {
+        public override void ConfigureBundle(BundleConfigurationContext context)
+        {
+            context
+                .Files
+                .AddIfNotContains("/Views/Settings/NotificationsSettingGroup/Default.js");
+        }
+    }
+
+    public class NotificationsSettingStyleBundleContributor : BundleContributor
+    {
+        public override void ConfigureBundle(BundleConfigurationContext context)
+        {
+           context.Files
+                .AddIfNotContains("/Views/Settings/NotificationsSettingGroup/Default.css");
+        }
     }
 }
