+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Bundling/UnityThemeUX2GlobalStyleContributor.cs
@@ -20,13 +20,9 @@ public class UnityThemeUX2GlobalStyleContributor : BundleContributor
         context.Files.AddIfNotContains("/libs/datatables.net-fixedheader-bs5/css/fixedHeader.bootstrap5.css");
         context.Files.AddIfNotContains("/libs/datatables.net-staterestore-dt/css/stateRestore.dataTables.css");
         context.Files.AddIfNotContains("/libs/tributejs/dist/tribute.css");
-        context.Files.AddIfNotContains("/libs/tinymce/skins/ui/oxide/content.css");
-        context.Files.AddIfNotContains("/libs/tinymce/skins/content/default/content.css");
-        context.Files.AddIfNotContains("/libs/tinymce/skins/ui/oxide/skin.css");
 
         // Add assets for "/themes/ux2/fonts/**/*"
-        context.Files.AddRange(new List<BundleFile> {
-            "/themes/ux2/fonts/icons/Segoe-Fluent-Icons.ttf",
+        context.Files.AddRange(["/themes/ux2/fonts/icons/Segoe-Fluent-Icons.ttf",
             "/themes/ux2/fonts/icons/Segoe-MDL2-Assets.ttf",
             "/themes/ux2/fonts/BCSans/BCSans-Bold.otf",
             "/themes/ux2/fonts/BCSans/BCSans-Bold.ttf",
@@ -52,6 +48,6 @@ public class UnityThemeUX2GlobalStyleContributor : BundleContributor
             "/themes/ux2/fonts/BCSans/BCSans-Regular.ttf",
             "/themes/ux2/fonts/BCSans/BCSans-Regular.woff",
             "/themes/ux2/fonts/BCSans/BCSans-Regular.woff2"
-        });
+        ]);
     }
 }
