+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/UX2/Layouts/Application.cshtml
@@ -56,7 +56,8 @@
     <title>@pageTitle</title>
 
     <abp-style-bundle name="@UnityThemeUX2Bundles.Styles.Global" />
-
+    <abp-style-bundle name="@NotificationsThemeBundles.Styles.Notifications" />
+    
     @await Component.InvokeAsync(typeof(WidgetStylesViewComponent))
 
     @await RenderSectionAsync("styles", false)
@@ -106,7 +107,7 @@
         @await Component.InvokeLayoutHookAsync(LayoutHooks.PageContent.Last, StandardLayouts.Application)
     </div>
 
-    <abp-script-bundle name="@UnityThemeUX2Bundles.Scripts.Global" />
+    <abp-script-bundle name="@UnityThemeUX2Bundles.Scripts.Global" />    
 
     <script src="~/Abp/ApplicationLocalizationScript?cultureName=@CultureInfo.CurrentUICulture.Name"></script>
     <script src="~/Abp/ApplicationConfigurationScript"></script>
