+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/UnityAspNetCoreMvcUIThemeUX2Module.cs
@@ -57,7 +57,7 @@ public class UnityAspNetCoreMvcUIThemeUX2Module : AbpModule
                 {
                     bundle
                         .AddBaseBundles(StandardBundles.Styles.Global)
-                        .AddContributors(typeof(UnityThemeUX2GlobalStyleContributor));
+                        .AddContributors(typeof(UnityThemeUX2GlobalStyleContributor));                        
                 });
 
             options
